Replace last int by long, solves the scheduling of medium size SIFT

+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/evaluator/SDFDoubleEdgePropertyType.java
@@ -67,7 +67,7 @@ public class SDFDoubleEdgePropertyType extends AbstractEdgePropertyType<Double> 
 
   @Override
   public long longValue() {
-    return this.value.intValue();
+    return this.value.longValue();
   }
 
   /*
